알람A 테스트

+++ clock/Drivers/STM32F4xx_HAL_Driver/Inc/stm32f4xx_hal_rtc.h
@@ -83,14 +83,14 @@ typedef struct
   */
 typedef struct
 {
-  int Hours;            /*!< Specifies the RTC Time Hour.
+  uint8_t Hours;            /*!< Specifies the RTC Time Hour.
                                  This parameter must be a number between Min_Data = 0 and Max_Data = 12 if the RTC_HourFormat_12 is selected
                                  This parameter must be a number between Min_Data = 0 and Max_Data = 23 if the RTC_HourFormat_24 is selected */
 
-  int Minutes;          /*!< Specifies the RTC Time Minutes.
+  uint8_t Minutes;          /*!< Specifies the RTC Time Minutes.
                                  This parameter must be a number between Min_Data = 0 and Max_Data = 59 */
 
-  int Seconds;          /*!< Specifies the RTC Time Seconds.
+  uint8_t Seconds;          /*!< Specifies the RTC Time Seconds.
                                  This parameter must be a number between Min_Data = 0 and Max_Data = 59 */
 
   uint8_t TimeFormat;       /*!< Specifies the RTC AM/PM Time.
